chore: rename to `poetry-plugin-bundle`
