+++ src/poetry_plugin_bundle/bundlers/bundler.py
